# Improve performance of Memory<T>.Span property getter