… src/System.Private.CoreLib/shared/System/Runtime/InteropServices/MemoryMarshal.cs
… tests/CoreFX/CoreFX.issues.json