Perf results for accessing the `Memory<T>.Span` property getter given different backing objects:

|                      Method |        Toolchain |        Mean |      Error |     StdDev | Scaled |
|---------------------------- |----------------- |------------:|-----------:|-----------:|-------:|
|              CharFromString | baseline (27008) |  3,360.8 ns |  31.849 ns |  28.233 ns |   1.00 |
|              CharFromString |       w/ changes |  2,946.9 ns |  24.783 ns |  21.969 ns |   0.88 |
|                             |                  |             |            |            |        |
|         CharFromArrayOfChar | baseline (27008) |  9,092.7 ns | 156.120 ns | 146.035 ns |   1.00 |
|         CharFromArrayOfChar |       w/ changes |  3,551.1 ns |  40.715 ns |  36.093 ns |   0.39 |
|                             |                  |             |            |            |        |
| CharFromMemoryManagerOfChar | baseline (27008) |  9,088.1 ns |  83.590 ns |  78.190 ns |   1.00 |
| CharFromMemoryManagerOfChar |       w/ changes |  6,460.6 ns |  36.038 ns |  28.136 ns |   0.71 |
|                             |                  |             |            |            |        |
|         ByteFromArrayOfByte | baseline (27008) |  8,479.8 ns | 219.163 ns | 277.172 ns |   1.00 |
|         ByteFromArrayOfByte |       w/ changes |  3,233.9 ns |  30.808 ns |  27.310 ns |   0.38 |
|                             |                  |             |            |            |        |
|        ByteFromArrayOfSByte | baseline (27008) | 42,603.5 ns | 685.866 ns | 641.559 ns |   1.00 |
|        ByteFromArrayOfSByte |       w/ changes |  3,381.6 ns |  71.574 ns | 155.597 ns |   0.08 |
|                             |                  |             |            |            |        |
| ByteFromMemoryManagerOfByte | baseline (27008) |  8,816.3 ns | 175.853 ns | 164.493 ns |   1.00 |
| ByteFromMemoryManagerOfByte |       w/ changes |  6,569.5 ns |  35.088 ns |  31.105 ns |   0.75 |
|                             |                  |             |            |            |        |
|      GetSpanFromEmptyMemory | baseline (27008) |  1,207.4 ns |  22.214 ns |  19.693 ns |   1.00 |
|      GetSpanFromEmptyMemory |       w/ changes |    887.2 ns |   4.268 ns |   3.784 ns |   0.74 |

Various optimizations include:

* We can rely on the `Memory<T>` ctor and factory methods to ensure that an object of an unexpected type never makes it in to the `_object` backing field, even in the face of a torn struct. This allows us to use unsafe code, bypassing the runtime type checks, but still requires that we perform bounds checks as appropriate.

* Single exit from the _Span_ property getter allows the JIT to optimize stack usage and minimize unnecessary data copying.

* Since we're part of coreclr, we can use deep knowledge of runtime object layout and method table layout to further skip some type checks. This is particularly useful in the case where an array is the backing store.

* The _Length_ property getter is once again just a simple field access with no bitmask logic.

* The _Span_ property getter's code gen size shrinks by around 60%.

Unit tests will not pass until the corresponding corefx change comes online since the unit tests perform private reflection over the `Memory<T>` backing fields.

I realize that this may be contentious, especially considering the fact that this proposes probing into the internals of how objects are represented within the runtime. But coreclr code is uniquely positioned to take advantage of these implementation details since the managed code can evolve in sync with any VM changes.

I'm also open to any recommendations people might have for real-world benchmarks so that we can test whether these changes are actually useful in practice.