@GrabYourPitchforks 2018-10-23
The GetHashCode changes in the latest iteration aren't needed since we should never be able to construct a `Memory<T>` with a null backing object and a non-zero _index_ / _length_. I'll revert them.