@jkotas 2018-10-25
I am ok with depending on internals like this as long as it is within CoreLib.

I am a bit worried about subtle GC holes that `GetObjectMethodTablePointer` may introduce in certain situations.