@GrabYourPitchforks 2018-10-25
@jkotas You mentioned a potential GC hole in `GetObjectMethodTablePointer`. Can you elaborate? I _believe_ this should be GC-safe, since getting the `ref int` to the first field will be an interior managed pointer back into the object, which is GC-tracked. If the object doesn't have any data fields, then the managed pointer should point _just past_ the end of the object, which the GC treats as equivalently to an interior pointer into the object itself. Then backing up by sizeof(void*) should backtrack no earlier than the original object reference pointer, which is still GC-tracked.

One of my assumptions above may be incorrect. If so we can change the method to be more reliable.