@GrabYourPitchforks 2018-11-02
Rebased PR on top of latest master to pick up Jan's earlier changes re: suppressing failing unit tests.