@GrabYourPitchforks 2018-11-06
Rebased on top of latest master due to last week's `Memory` / `ReadOnlyMemory` changes causing conflicts. Latest iteration also addresses the remaining comment cleanup, further improves `Span` property getter performance (using the slice trick we committed last week), and changes `GetHashCode` to be simpler but subject to a higher performance hit.